book: merge certificate code files
