+++ docs/book/Cargo.toml
@@ -21,10 +21,6 @@ name = "certificate-insecure"
 path = "src/quinn/certificate-insecure.rs"
 
 [[bin]]
-name = "certificate-certsr"
-path = "src/quinn/certificate-certs.rs"
-
-[[bin]]
 name = "data-transfer"
 path = "src/quinn/data-transfer.rs"
